remove util.js, change to utilities.ts, refactor imports
