+++ app/assets/javascripts/components/annotations/code_listing_row.ts
@@ -6,7 +6,7 @@ import "components/annotations/annotations_cell";
 import "components/annotations/annotation_marker";
 import "components/annotations/hidden_annotations_dot";
 import { i18nMixin } from "components/meta/i18n_mixin";
-import { initTooltips } from "util.js";
+import { initTooltips } from "utilities";
 import { PropertyValues } from "@lit/reactive-element";
 import { userState } from "state/Users";
 import { AnnotationData, annotationState, compareAnnotationOrders } from "state/Annotations";
