+++ app/assets/javascripts/institution.ts
@@ -1,4 +1,4 @@
-import { fetch } from "./util.js";
+import { fetch } from "utilities";
 
 function initInstitutionSelectionTable(institutionId: number): void {
     const institutionPanel = document.querySelector("#institution-panel");
