+++ app/assets/javascripts/repository.ts
@@ -1,4 +1,4 @@
-import { fetch } from "./util.js";
+import { fetch } from "utilities";
 import { Toast } from "./toast";
 
 function initAdminsEdit(): void {
