+++ app/assets/javascripts/visualisations/series_graph.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line
 // @ts-nocheck
 import * as d3 from "d3";
-import { initDatePicker } from "util.js";
+import { initDatePicker } from "utilities";
 
 export type RawData = {
     // eslint-disable-next-line camelcase
