+++ app/javascript/packs/announcement.js
@@ -1,4 +1,4 @@
-import { initDatePicker } from "util.js";
+import { initDatePicker } from "utilities.ts";
 import "components/datalist_input";
 
 dodona.initDatePicker = initDatePicker;
