+++ app/javascript/packs/evaluation.js
@@ -1,6 +1,6 @@
 import { initCheckboxes, initCheckbox, initEvaluationStepper } from "evaluation.ts";
 import FeedbackActions from "feedback/actions";
-import { initDatePicker } from "util.js";
+import { initDatePicker } from "utilities.ts";
 
 window.dodona.initDeadlinePicker = initDatePicker;
 window.dodona.initCheckbox = initCheckbox;
